functions/register_monitoring_data/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.3.2
google-cloud-firestore==2.5.3